% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{add.m}
\alias{add.m}
\title{Addition modulo m. Defaults to Boolean arithmetic XOR.}
\usage{
add.m(x,2) returns the sum mod 2 of the vector x
add.m(x,y,2) returns the (x + y) mod 2.
}
\arguments{
\item{x}{a number or a vector}

\item{y}{a number, defaults to NULL}

\item{m}{the modulus, defaults to 2 for binary addition (XOR)}
}
\description{
A function to do addition modulo some integer, m.
}
