% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_eq.R
\name{affine.map}
\alias{affine.map}
\title{Function for simple affine map}
\usage{
affine.map(x, a = 0.75, b = 1)
}
\description{
Facilitates use of the iterate.map() function for
studying behavior of simple affine maps; alternative to delta.eq() function
}
