% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{bernoulli.shift}
\alias{bernoulli.shift}
\title{Bernoulli shift map}
\usage{
bernoulli.shift(x)
}
\arguments{
\item{x}{last value of x, i.e. x[n]}
}
\description{
funny little map, demonstratively useful, otherwise...?
}
\examples{
bernoulli.shift(.25)
bernoulli.shift(.5)
bernoulli.shift(2)
}
\keyword{bernoulli.shift}
