% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bifurcation.R
\name{bifurcation}
\alias{bifurcation}
\title{Plot a bifurcation diagram}
\usage{
bifurcation(
  parmin = 1,
  parmax = 4,
  f = logistic.eq,
  x0 = 0.3,
  niter = 100,
  length = 301,
  usepar = FALSE,
  xmin = 0,
  ylim = c(-0.1, 1),
  cex = 0.1,
  pch = 19,
  xlab = "Parameter value",
  ylab = expression(x[n])
)
}
\arguments{
\item{parmin}{minimum parameter value, defaults to 1}

\item{parmax}{maximum parameter value, defaults to 4}

\item{f}{function defining the map, defaults to logistic equation}

\item{x0}{initial value to begin trajectories from (constant over parameter
variations)}

\item{niter}{number of iterations to assess "long-term" behavior of the map}

\item{length}{length of the vector of parameter values to be assessed}

\item{usepar}{if TRUE, use maximum of parmin or xmin for the lower limit of xlim
in the underlying call to plot()}
}
\description{
Currently only works for one parameter models. Plots the long-term
behavior of a dynamical system as a function of varying its parameter value.
}
\examples{
bifurcation() #produces the logistic map's bifurcation diagram
bifurcation(f=gauss4bifurc, parmin=-1, parmax=1, ylim=c(-1,1.5),
usepar=TRUE, xmin=-1, length=201, x0=.2)
}
