% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{binary}
\alias{binary}
\title{Convert decimal integers to binary representation.}
\usage{
binary(i, type = "vector")
}
\arguments{
\item{i}{an interger-valued number}

\item{type}{= ("vector", "binary"): should the results be returned
as a vector of binary digits or as a single binary number? Defaults
to vector}
}
\description{
Based on a function by Thomas Lumley
}
\examples{
binary(90)
binary(150)
binary(150, type="binary")
}
