% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{boolean.update}
\alias{boolean.update}
\title{Small example Boolean network}
\usage{
boolean.update(x)
}
\arguments{
\item{x}{an initial state vector for the 4 node network}
}
\description{
From Veliz-Cuba (2010) a four node Boolean network

Note: This is not a Kauffman network, as nodes 1 and 2 have two inputs
each, while nodes 3 and 4 have one input each.

x_1 = x_2^x_3

x_2 = x_1^-x_4

x_3 = x_2

x_4 = -x_3
}
\examples{
boolean.update(c(0,1,1,0))
}
