% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{ca.initialize}
\alias{ca.initialize}
\title{Initialize a vector for initial condition of a cellular automata simulation}
\usage{
ca.initialize(L, type = c("middle", "random"))
}
\arguments{
\item{L}{the length of the desired vector; corresponds to the size
of the one-dimensional lattice the cellular automata will "live" on}

\item{type}{which type of initialization, "middle" will set the center
cell to 1, all else to zero; "random" sets the lattice to a random configuration
with any cell equal to 1 with probability .5}
}
\description{
Initializes a vector of a given size, with only one cell (the "middle") at 1,
all others 0; initial condition of the 1-D lattice.
}
