% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{ca.plot}
\alias{ca.plot}
\title{Space-time plot}
\usage{
ca.plot(ca, col = c("white", "black"))
}
\arguments{
\item{ca}{A matrix containing the space-time evolution for a 1D CA model}

\item{col}{Colours for the space-time plot (defaults to white for "off" and
black for "on")}
}
\description{
Generate the space time plot for 1D cellular automata models.
Useful for other applications, like percolation theory.
just a wrapper for the image function that calls the rotate.mat
function above and sets the margins neatly.
}
