% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{catalan}
\alias{catalan}
\title{Catalan numbers}
\usage{
catalan(n)
}
\arguments{
\item{n}{index of the Catalan number, e.g., n = 1, 1; n = 2, 2, n = 3, 5, ... etc.}
}
\description{
calculate the nth Catalan number
}
\examples{
catalan(n=4)
}
