% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{closed.form2}
\alias{closed.form2}
\title{Closed form solution for second order difference equation
initial value problem}
\usage{
closed.form2(root1, root2, y0, y1, echo = TRUE)
}
\arguments{
\item{root1}{First root of the characteristic equation (from quad.roots)}

\item{root2}{Second root of the characteristic equation}

\item{y0}{time 0 initial value}

\item{y1}{time 1 initial value}

\item{echo}{defaults to TRUE, when TRUE,
prints  the closed form solution}
}
\value{
A list with the closed formula solution as a function of k,
constant c1, and constant c2
}
\description{
Find solution to a second order difference equation:
s(k) = c1*m_1^k + c2*m_2^k, where m_1 and m_2 are the roots of the
characteristic equation for the second order difference equation.
}
