% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{cobweb}
\alias{cobweb}
\title{Cobweb analysis of a 1-dimensional dynamical system}
\usage{
cobweb(x, type = "segments", start.y = 0, start.x = 0, col = "blue", lwd = 2)
}
\arguments{
\item{x}{a vector of the values of the sequence given by the difference}

\item{type}{defaults to segments, can also choose "arrows"}

\item{start.y}{initial position on y-axis, defaults to 0}
}
\description{
"cobweb" a model (step through iterations); also called
"graphical iteration" (e.g., Sternberg, 2010).
}
\examples{
cobweb()
}
\keyword{cobweb}
