% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{decimal}
\alias{decimal}
\title{Converts a binary vector to its decimal equivalent}
\usage{
decimal(bin)
}
\arguments{
\item{bin}{A vector of length N, containing entries from the set {0,1}
with the least significant digit last (i.e., reads left to right)}
}
\description{
useful for calculating the Wolfram Code of a CA rule
}
