% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_eq.R
\name{delta.eq}
\alias{delta.eq}
\title{Simulate from 1st order linear (affine) difference
Equations}
\usage{
delta.eq(y0 = 1, a = 2, b = 0, N = 5, monitor = FALSE)
}
\arguments{
\item{y0}{initial value for sequence, defaults to 1}

\item{a}{slope parameter for linear model, defaults to 2}

\item{b}{constant parameter for linear model, defualts to 0}

\item{N}{length of sequence, forward iterations = N-1, defaults to 5}

\item{monitor, }{print iteration number and next value? defaults to FALSE}
}
\description{
Difference equations of this form can correspond to either
arithmetic growth or geometric growth, depending on their
pattern of parameters (and initial values).
for A = 1, B = b: arithmetic growth
for A = a, B = 0: geometric growth (or, really, B = b)
}
\examples{
x <- delta.eq()
series.plot(x)
}
\keyword{difference}
\keyword{equation}
