% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_eq.R
\name{delta.two}
\alias{delta.two}
\title{Second-order linear autonomous difference equation}
\usage{
delta.two(y0 = 2, y1 = 3, a = 1, b = -0.5, c = 1, N = 25, monitor = FALSE)
}
\arguments{
\item{y0}{first intial value, defaults to 2}

\item{y1}{second initial value, defaults to 3}

\item{a}{coefficient on "next" time step of y (y[n+1]), defaults to 1}

\item{b}{coefficient on current time step of y (y[n]), defaults to -0.5}

\item{c}{constant, defaults to 1}

\item{N}{number of iterations (includes first two initial values)}

\item{monitor}{print iteration number and next step value of y, defaults to FALSE}
}
\description{
defaults to Harrod-Domar-Hicks model of the growth of
national income (oscillatory convergence version)
as described in S. Goldberg (1958) "Introduction to difference
Equations" Chapter 0.
Takes 2 initial values, as needed to define a second-order recurrence
}
\examples{
y <- delta.two()
series.plot(y)
# The Fibonacci sequence is a 2nd-order recurrence, with these settings
F <- delta.two(y0=0,y1=1, a=1, b=1, c=0, N=11)
F
series.plot(F)
}
\keyword{difference}
\keyword{equation}
\keyword{order}
\keyword{second}
