% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{discrete.SIR}
\alias{discrete.SIR}
\title{Discrete time SIR epidemiological model of infectious disease}
\usage{
discrete.SIR(states, beta, alpha, N = 1)
}
\arguments{
\item{states}{the value of the S,I, and R compartments: should be done as percentage of population;
so initial S should be 1, common starting points for I might be .01 or .1, and R should be 0.}

\item{beta}{the transmission rate of the disease}

\item{alpha}{the rate defining transition from the I to the R compartment.}

\item{N}{the total population size to scale the beta coefficient, defaults to 1 so as to not
do any such scaling.}
}
\description{
A basic Susceptibles-Infectives-Removeds epidemiological model
for infectious disease. This is a compartment model defined by three equations
which describe the transfer of individuals between compartments: Susceptibles are
capable of becoming infected and enter the infectives compartment when they do.
Infectives enter the "removed" compartment by recovering or dying (both
are captured in a single rate parameter in this implementation.)
}
