% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{euler}
\alias{euler}
\title{Basic numerical integration by Euler's method}
\usage{
euler(y0 = 1, f, dx = 0.01, nSteps = 1000)
}
\arguments{
\item{y0}{starting value for y defaults to 1}

\item{f}{the function to be integrated over}

\item{dx}{step size for numerical integration, defaults to .01}

\item{nSteps}{number of integration steps, defaults to 1000}
}
\description{
Numerically solve a single OBE by numerical integration
using Euler's method:
dy/dx = f(x)
(y[n+1] - y[n])/dx = f(x) for small dx
y[n+1] = y[n] + dx*f(x)
}
\examples{
f <- function(x) 2.75*x*(1-x)
init <- runif(1)
sol <- euler(init, f=f)
plot(y ~ times, data=sol, type="l")
}
