% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{every.state}
\alias{every.state}
\title{Every state for a Boolean system of order K}
\usage{
every.state(order)
}
\description{
This function produces a matrix whose rows
contain every possible state for a Boolean system with K nodes.
}
\examples{
every.state(2)
every.state(4)
}
