% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{fib.triangle}
\alias{fib.triangle}
\title{Uses the Fibonacci sequence as the basis for triangle number summation.}
\usage{
fib.triangle(N)
}
\arguments{
\item{N}{the final Fibonacci number to sum to}
}
\description{
Uses the Fibonacci sequence as the basis for triangle number summation.
}
