% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{fibonacci}
\alias{fibonacci}
\title{Return a sequence of Fibonacci numbers}
\usage{
fibonacci(N, y0 = 0, y1 = 1, monitor = FALSE)
}
\arguments{
\item{N}{the last Fibonacci number in the sequence}

\item{y0}{the first initial value, defaults to 0}

\item{y1}{the second initial value, defaults to 1}

\item{monitor}{when TRUE prints the iterations starting from the second Fibonacci number. Defaults to
FALSE.}
}
\description{
Get the sequence of Fibonacci numbers to N. Defaults to the Fibonacci sequence starting
from 0: 0,1,1,2,3,5,8,13,21,...
Initial values default to 0 and 1, so the length of the sequence will be N+1. If the user provides
a starting value for y0 of 1, then the resulting sequence will be of length N. Both initial values
are user-settable, allowing exploration of any Fibonacci-type sequence. The function is a masked call to
the delta.two function, so users interested in more elaborate sequence control can use that to set the
parameters (the fibonacci function hard-sets them: a = 1, b = 1, c = 0).
}
\examples{
F <- fibonacci(N=15, monitor=TRUE)
F
}
