% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{fill}
\alias{fill}
\title{Fill a vector with leading zeros}
\usage{
fill(x, a)
}
\arguments{
\item{x}{an initialized vector (can be set to all NA) of length > length(a)}

\item{a}{the content (typically a vector with length less than length(x))}
}
\description{
Add leading zeroes ahead of content to a vector
}
