% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{find.order}
\alias{find.order}
\title{Find the order of a polynomial equation generating a sequence of numbers}
\usage{
find.order(vec, tol = 0.01, print = FALSE, diff.table = FALSE)
}
\arguments{
\item{vec}{the input sequence}

\item{tol}{tolerance for determining whether an order of difference is "constant". Defaults to .01.}

\item{print}{logical, print the differenced sequences for each iteration, defaults to FALSE}

\item{diff.table}{logical, include a table of differences in output, defaults to FALSE}
}
\description{
The function differences the observed sequence until the differenced sequence is
constant and returns the order of differencing and the value of the constant. For instance, if
the order of differencing returned is 2, then the sequence can be generated using a second order
polynomial. Constancy is determined by assessing whether the range of the difference
sequence is below a tolerance threshold (can be set to 0 for absolute constancy).
}
\examples{
x <- c(1,3,6,10,15,21,28)
find.order(x)
}
