% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{first.solve}
\alias{first.solve}
\title{Solution to a first-order recurrence/difference equation}
\usage{
first.solve(x0, model = NULL, r = NULL, b = NULL)
}
\arguments{
\item{x0}{the initial condition for the system}

\item{model}{a least squares fit for the system (an lm object)}

\item{r}{as above}

\item{b}{as above}
}
\description{
calculates solution  to

y(n + 1) = r*y(n) + b

in the form

y(k) = r^k*(y0 - y.star) + y.star

where y.star is the equilibrium, b/(1 - r).
}
