% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{fixed.pt}
\alias{fixed.pt}
\title{Calculate steady state/equilibrium of a linear dyamical system}
\usage{
fixed.pt(model = NULL, a = NULL, b = NULL)
}
\arguments{
\item{model}{a fitted model of class 'lm', defaults to NULL}

\item{a}{slope parameter of a linear recursion relation}

\item{b}{constant parameter of a linear recursion relation}
}
\description{
fixed.pt() calculates the fixed point of a
first order, autonomous linear difference equation model.
Input can be either an lm-class model (model) OR a & b parameters
of a theoretical model
}
\examples{
fixed.pt(a=2,b=1)
}
\keyword{equilibrium,}
\keyword{fixed}
\keyword{point}
