% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{fixpt.2}
\alias{fixpt.2}
\title{Fixed point of a second order difference equation}
\usage{
fixpt.2(model = NULL, a, b, c)
}
\arguments{
\item{model}{a fitted model of type lm defaults to NULL}

\item{a}{coefficient on y[n+1]}

\item{b}{coefficient on y[n]}

\item{c}{constant}
}
\description{
Calculate the fixed point/equilibrium/steady state for
a 2nd order model of the form y[n+2] = a*y[n+1] + b*y[n] + c as
c/(1 - a - b). Can calculate from either a fitted model from lm or
user-specified parameters from a theoretical model.
}
