% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{gauss.map}
\alias{gauss.map}
\title{The Gauss map}
\usage{
gauss.map(x, a = 4.9, b = -0.58)
}
\arguments{
\item{x}{variable to iterate}

\item{a}{the a parameter, defaults to 4.9}

\item{b}{the b parameter, defaults to -0.58}
}
\description{
x = exp(-a*x^2) + b \\
The Gauss iterated map is sometimes called the "mouse" map
because its bifurcation diagram looks like a mouse
}
