% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{gauss4bifurc}
\alias{gauss4bifurc}
\title{Special version of the Gauss map for bifurcation plotting}
\usage{
gauss4bifurc(x, b)
}
\description{
gauss4bifurc holds the a-parameter of the Gauss map
constant at 4.9, while allowing the b-parameter to vary to produce
the "mouse" map.
}
\examples{
bifurcation(f=gauss4bifurc, parmin=-1, parmax=1, ylim=c(-1,1.5),
usepar=TRUE, xmin=-1, length=201, x0=.2)
}
