% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic.R
\name{harmonic}
\alias{harmonic}
\title{Calculate the first through Nth Harmonic Numbers}
\usage{
harmonic(N = 40)
}
\arguments{
\item{N}{which Harmonic number do you want to get up to?}
}
\description{
Calculate the first through Nth Harmonic Numbers
}
\examples{
harmonic()
}
