% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henon.R
\name{henon.plot}
\alias{henon.plot}
\title{Plot the output of the Henon Map function}
\usage{
henon.plot(h)
}
\arguments{
\item{h}{a henon object}
}
\description{
Plot the output of the Henon Map function
}
\examples{
out <- henon()
henon.plot(out)
}
