% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/henon.R
\name{henon}
\alias{henon}
\title{Henon map (a 2-D map displaying chaos)}
\usage{
henon(x = 0, y = 0, a = 1.4, b = 0.3, nSteps = 10000)
}
\arguments{
\item{x}{x-coordinate, defaults to 0}

\item{y}{y-coordinate, defaults to 0}

\item{a}{the a parameter (coefficient on x[n]^2 in the x equation), default is 1.4}

\item{b}{the b parameter (coefficient on x[n] in the y equation), default is 0.3}

\item{nSteps}{the number of iterations, default is 10000}
}
\description{
Described in Henon, 1976. Default values come from this paper.
x = 1 - a*x^2 + y # "fold"
y = b*x           # "flip" and contract
}
\examples{
out <- henon()
henon.plot(out)
}
