% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{identity}
\alias{identity}
\title{Identity function}
\usage{
identity(x)
}
\arguments{
\item{x}{identity function gives f(x) = x}
}
\description{
useful utility for plotting cobweb analyses
identifies fixed points when identity(x) = map(x)
}
\keyword{identity}
