% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-automata.R
\name{iterate.1dca}
\alias{iterate.1dca}
\title{iterate a 1D Cellular automaton (or Boolean network or system of equations)}
\usage{
iterate.1dca(x0, f = R170, nSteps = 10)
}
\arguments{
\item{x0}{initial state of the system}

\item{f}{dynamical rule or "updating function". Defaults to Wolfram rule 170.}

\item{nSteps}{number of time steps to iterate the simulation forward;
defaults to 10}
}
\description{
This function will take a dynamical rule and an initialized vector
and iterate it forward in time for nSteps iterations
}
\examples{
x0 <- initialization(10)
x <- iterate.1dca(x)
}
