% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{iterate.2dca}
\alias{iterate.2dca}
\title{iterate a 2-dimensional cellular automaton model}
\usage{
iterate.2dca(x, f = life, nSteps = 10)
}
\arguments{
\item{x}{the initial state matrix for the model}

\item{f}{the function defining the system's updates, the system's dynamical rule, defaults to 'life'}

\item{nSteps}{the number of time steps to iterate the system for}
}
\description{
Takes a prepared lattice of cells and iterates a dynamical rule/update function to
produce a spacetime array for the lattice model. Outputs a three way array, such that the rows and columns
the cells of the lattice at a given time step such that each matrix slice contains the system state at
a given time point/iteration.
}
