% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{iterate.map}
\alias{iterate.map}
\title{Iterate a given user defined 1-dimensional map}
\usage{
iterate.map(a0, f = a, nSteps = 11, ...)
}
\arguments{
\item{a0}{initial value of the sequence to be iterated}

\item{f}{"next value" function (f, which defaults to a function called "a")}

\item{nSteps}{the number of steps to iterate (include a0, forward iterations
equal to nSteps-1)}
}
\description{
Allows modular examination of 1-dimensional discrete dynamical systems
}
\examples{
iterate.map()
}
\keyword{iterate}
\keyword{update,}
