% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{life}
\alias{life}
\title{update function for Conway's Game of Life}
\usage{
life(L)
}
\arguments{
\item{L}{a lattice (typically provided by the make.life function)}
}
\description{
Assumes a stable boundary that remains zero. Within this boundary, uses the Moore neighbourhood
around each cell to determine birth, life, and death:
(a) if a cell is currently occupied and two or three of its neighbouring cells are also occupied, the cell
remains alive, (b) if the cell is currently unoccupied and exactly three of its neighbours are occupied a
new "occupant" is born in the cell, and (c) if the cell is occupied and more than three of its 
neighouring cells are occupied, the occupant 'dies' - if the cell is unoccupied under this condition, it remains
empty.
}
