% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor-logistic.R
\name{lin.coef}
\alias{lin.coef}
\title{calculations for linearizing the logistic model}
\usage{
lin.coef(mu)
}
\description{
Private function for calculating the coefficients
of a first-order Taylor series approximation of the logistic
equation
}
