% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{lin.fit}
\alias{lin.fit}
\title{Fit a linear model to a first-order difference equation}
\usage{
lin.fit(vec, center = FALSE)
}
\arguments{
\item{vec}{the vector of data}

\item{center}{set the first observation to zero and difference all observations from it, defaults to zero}
}
\value{
the fit of the model, an object of class 'lm'
}
\description{
Function takes a vector of observed time-ordered data and fits a linear model for its
first order difference model
}
