% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylor-logistic.R
\name{lin.logistic}
\alias{lin.logistic}
\title{Taylor Series linearization of the logistic equation}
\usage{
lin.logistic(y, mu)
}
\arguments{
\item{y}{a value of the variable to be iterated}

\item{mu}{the parameter of the logistic model}
}
\description{
Approximates the logistic model near its equilibrium
}
\examples{
lin.logistic(y=.62,mu=2.5)
# for comparison
logistic.eq(x=.62, mu=2.5)
}
\keyword{approximation}
