% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_eq.R
\name{linear.map}
\alias{linear.map}
\title{Function for simple linear map}
\usage{
linear.map(x, a = 0.5)
}
\description{
. Facilitates use of the iterate.map() function for
studying behavior of simple linear maps; alternative to delta.eq() function.
Wrapper for the affine.map() function with b set to 0.
}
