% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{logistic.eq}
\alias{logistic.eq}
\title{Implements the logistic equation}
\usage{
logistic.eq(x, mu = 2.5)
}
\arguments{
\item{x}{variable to iterate on}

\item{mu}{parameter of the logistic map x[n+1] = a*x[n]*(1 - x[n]), defaults to 2.5}

\item{N}{number of iterations (includes initial value), defaults to 20}
}
\value{
for x = x[n], returns x[n+1]
}
\description{
For use with the iterate.map() function. Function provides the 'next
 time step' value in the sequence {x[n]}. That is, for x[n] -> x[n+1]
}
\examples{
logistic.eq(.5, a=2.5)
}
\keyword{logistic.eq}
