% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{logistic.fixedpt}
\alias{logistic.fixedpt}
\title{Calculate the fixed point of the logistic map:}
\usage{
logistic.fixedpt(mu)
}
\arguments{
\item{a}{parameter of the logistic map x[n+1] = mu*x[n]*(1 - x[n])}
}
\description{
Calculate the fixed point of the logistic map:
}
\examples{
logistic.fixedpt()
}
\keyword{equilibrium}
