% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{make.adjacency}
\alias{make.adjacency}
\title{Make the adjacency matrix for state transitions for a boolean network}
\usage{
make.adjacency(order = 4, system = boolean.update)
}
\description{
Uses the next.states function to determine the state transition for all possible states of
a Boolean network of a given order and a given dynamic/updating rule. Intended to provide input
for graphing the state transitions (using igraph)
}
\examples{
adjm <- make.adjacency()
g1 <- graph_from_adjacency_matrix (x)
plot(g1)
}
