% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life.R
\name{make.life}
\alias{make.life}
\title{Create lattice for Conway's Game of Life}
\usage{
make.life(k = 10, prob = 0.5)
}
\arguments{
\item{k}{the size of the lattice (makes a square lattice), defaults to 10}

\item{prob}{the probability that a cell is occupied, defaults to .5}
}
\description{
wrapper and extension of make.lattice. Appends cells to edges of the k x k lattice. The life
function assumes this sort of boundary (i.e., the finite lattice we simulate on isolated from the rest of
the theoretically infinite lattice by a stable boundary layer of cells that remain zero 'for all time').
}
