% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{make.plot}
\alias{make.plot}
\title{"Graph paper" plot}
\usage{
make.plot(xlim = c(-5, 5), ylim = c(-5, 5))
}
\arguments{
\item{xlim}{minimum and maximum values of the x-axis}

\item{ylim}{minimum and maximum values of the y-axis}
}
\description{
generates a plot space with gridlines, like graph paper
}
