% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{middle}
\alias{middle}
\title{Find the middle value of a vector of size L}
\usage{
middle(L)
}
\arguments{
\item{L}{length of the 1-D lattice}
}
\description{
A helper function for initialization function
}
