% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{multiply.m}
\alias{multiply.m}
\title{Boolean arithmetic AND (multiplication mod 2)}
\usage{
multiply.m(x, y = NULL, m = 2)
}
\arguments{
\item{x}{a number or a numeric vector}

\item{y}{a number, defaults to NULL}

\item{m}{the modulus, defaults to 2 for binary multiplication (AND)}
}
\description{
Defaults to Boolean arithmetic AND, but can be
used for general multiplication mod m
}
