% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{newton}
\alias{newton}
\title{Newton's method}
\usage{
newton(f, x0 = 2, iter = 100)
}
\arguments{
\item{f}{the function to find a root}

\item{x0}{the starting value to evaluate}

\item{iter}{number of iterations to carry out the algorithm}
}
\description{
Newton's method to iteratively find the roots of a function
For a function P(x), the algorithm is
x_[n+1] = x_n - (P(x_n)/P'(x_n)), where P'(x) is the first derivative
of P(x).
}
\examples{
f <- function(x) x^2 - 2
newton(f=f, x0= 99, iter=14)
}
