% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{next.states}
\alias{next.states}
\title{next state finder}
\usage{
next.states(order = 4, system = boolean.update)
}
\arguments{
\item{order}{the number of nodes in the Boolean system, defaults to 4}

\item{system}{the update rules for each node of the system, defaults to
'boolean.update'}
}
\description{
Find next state for any input on a Boolean system
}
