% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{or}
\alias{or}
\title{Boolean arithmetic OR. (addition mod 2)}
\usage{
or(x, y, m = 2)
}
\arguments{
\item{x}{input from the set {0,1}/default operation}

\item{y}{input from the set {0,1}}

\item{m}{modulus for the operation}
}
\description{
(x + y + x*y) mod 2
}
\examples{
or(0,0)
or(0,1)
or(1,0)
or(1,1)
}
