% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{orbit2D}
\alias{orbit2D}
\title{Plot a trajectory/orbit for a 2-D map}
\usage{
orbit2D(trajectory, lty = 1, pch = 19, lwd = 1, cex = 1)
}
\arguments{
\item{trajectory, }{a 2 column matrix of the trajectory of a 2D Dynamical system}
}
\description{
Plot a trajectory/orbit for a 2-D map
}
\examples{
plot.orbit2D()
}
\keyword{orbit,}
\keyword{trajectory}
