% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolation-functions.R
\name{perc.prob.1d}
\alias{perc.prob.1d}
\title{Proababilty a finite 1-d lattice will percolate}
\usage{
perc.prob.1d(L, p = 0.5)
}
\arguments{
\item{L}{an integer giving the length of the lattice or a vector of sites defining the lattice}

\item{p}{site occupation probability, defaults to .5}
}
\description{
For an infinite lattice, the only site occupation probability that will
percolate on the lattice is 1. Any probability less than 1 will leave a proportion of
sites unoccupied that is equal to p.

For a finite lattice, site occupation probabilities less than 1 can produce a lattice that percolates,
but the probability decreases with lattice length. For lattice length L and site occupation
probability p:

Pr(percolates) = p^L
}
\examples{
perc.prob.1d(10)
perc.prob.1d(10, .55)
}
