% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{prep.cobweb}
\alias{prep.cobweb}
\title{Prep a cobweb analysis (graphical iteration)}
\usage{
prep.cobweb(
  f = logistic.eq(x, mu = 2),
  from = 0,
  to = 1,
  lwd = 2,
  col = "darkgray",
  ylim = NULL,
  xlab = expression(x[n]),
  ylab = expression(x[n + 1])
)
}
\arguments{
\item{f}{user defined function (update function, map), defaults to logistic
equation with parameter a = 2}

\item{from}{mask for curve function, defaults to 0}

\item{xlab, }{defaults to expression(x[n])}

\item{ylab, }{defaults to expression(x[n+1])}
}
\description{
Prep a cobweb analysis (graphical iteration)
}
\examples{
prep.cobweb()
}
\keyword{cobweb}
