% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{prep.matrix}
\alias{prep.matrix}
\title{Prepare a space-time matrix}
\usage{
prep.matrix(x0, nSteps = 11)
}
\arguments{
\item{x0}{a vector containing the initial state for all locations}

\item{nSteps}{number of iterations, defaults to 11}
}
\description{
Prepare a matrix to store space-time evolution of
1D cellular automata, Boolean network, or multiple variable dynamical
system
}
