% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{quad.roots}
\alias{quad.roots}
\title{Find the real or complex roots of a quadratic equation}
\usage{
quad.roots(a, b, c)
}
\arguments{
\item{a}{the coefficient on the squared term}

\item{b}{the coefficient on the order-1 term}

\item{c}{the constant}
}
\description{
Find roots of the equation a*m^2 + b*m + c = 0
}
