% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-automata.R
\name{R170}
\alias{R170}
\title{Wolfram rule 170}
\usage{
R170(x)
}
\arguments{
\item{x}{an initialized space-time cellular automata matrix}
}
\description{
Rule 170
depends on initialize() and prep.matrix()
Acts as a left-shift register with periodic boundary condition
(i.e., the lattice is a one-dimensional ring)
}
