% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-automata.R
\name{R184}
\alias{R184}
\title{Wolfram rule 184}
\usage{
R184(x)
}
\description{
Long term behaviour of this rule varies qualitatively
as a function of its initial state (can display Class 1, 2, or 3 behaviour).

S(p,q,r): qr + (1 - q)p
}
