% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-automata.R
\name{R22}
\alias{R22}
\title{Wolfram Rule 22}
\usage{
R22(sigma)
}
\description{
The cell is active in the next timestep if exactly one
of it and its left and right nearest neighbours is active.
}
