% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cellular-automata.R
\name{R254}
\alias{R254}
\title{Wolfram rule 254}
\usage{
R254(x)
}
\description{
Only a fully empty neighbourhood returns an inactive cell,
so almost all initializations land in the 'all-on' attractor. The only
initial condition that does not is the 'all-zero' condition, which results
in the repulsive "all-off" attractor. That is, any perturbation will
eventually lead to all cells being active.

S(p,q,r): 1 - (1-p)(1-q)(1-r)
}
