% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{riemann.sum}
\alias{riemann.sum}
\title{Numerical integration by Riemann Sum}
\usage{
riemann.sum(f, a = 0, b = 1, nSteps = 1000, type = "left")
}
\arguments{
\item{f}{the function to integrate}

\item{a}{lower limit, defaults to 0}

\item{b}{upper limit of integration, defaults to 1}

\item{nSteps}{number of integration (delta-x) steps to take, defaults to 1000}

\item{type}{should the function be evaluated on the left or right
side, defaults to left}
}
\description{
Numerically integrate a function f using Riemann's method
of sums of rectangles. This gives the definite integral with lower
limit a and upper limit b.
}
