% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{rotate.mat}
\alias{rotate.mat}
\title{Rotate a Matrix for plotting}
\usage{
rotate.mat(mat)
}
\arguments{
\item{a}{matrix (often containing the space-time diagram of a cellular
automaton model)}
}
\description{
adapted from cape package by
Anna L. Tyler and Wei Lu and Justin J. Hendrick and
Vivek M. Philip and Greg W. Carter
vectorized with help from SNAP Technologies at UAlaska-Fairbanks.
Primarily a utility function for plot.ca()
}
\keyword{utility}
