% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{runs.test}
\alias{runs.test}
\title{Runs test for randomness in a set of discrete values (especially binary)}
\usage{
runs.test(y)
}
\arguments{
\item{y}{a sequence of (binary) data}
}
\description{
Runs test for randomness in a set of discrete values (especially binary)
}
