% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{series.plot}
\alias{series.plot}
\title{Plot the time series of a one-dimensional map}
\usage{
series.plot(
  x,
  start = 1,
  xlab = "Time period n",
  ylab = expression(x[n]),
  type = "b",
  pch = 19,
  lwd = 1,
  lty = 1,
  col = "black",
  main = ""
)
}
\arguments{
\item{x}{sequence given by the difference equation}

\item{start}{what value to start x-axis at, defaults to 1}
}
\description{
x axis starts at value given by "start", ends at start + (length(x)-1)
}
\examples{
plot.series()
}
\keyword{orbit,}
\keyword{trajectory}
