% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{sim.chisq}
\alias{sim.chisq}
\title{Calculate chi-square of a forward simulated empirical model}
\usage{
sim.chisq(data, sim)
}
\arguments{
\item{data}{observed sequence of real values}

\item{sim}{the forward simulation sequence}
}
\description{
Chi-squared test statistic calculator for fit of
"forward simulated" model against observed data
}
\keyword{chisq}
