% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percolation-functions.R
\name{site.perc}
\alias{site.perc}
\title{Site percolation}
\usage{
site.perc(lattice, prob)
}
\arguments{
\item{lattice}{an initialized 1- or 2-dimensional lattice}

\item{prob}{probability that a cell is occupied}
}
\description{
fills the cells of a 1- or 2-dimensional square lattice
with probability given by 'prob'
}
