% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{states.plot}
\alias{states.plot}
\title{mask for the igraph package's plot function}
\usage{
states.plot(g, ...)
}
\description{
provides ease-of-use for those unfamiliar with the igraph packe to easily plot state transition
graphs for finite state dynamical systems
}
