% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taylorize.R
\name{taylorize}
\alias{taylorize}
\title{Taylorize an arbitrary one-dimesional map}
\usage{
taylorize(f, fixed.pt)
}
\description{
Linearize a nonlinear one-dimensional map using
a Taylor series expansion
}
