% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative-maps.R
\name{tent.eq}
\alias{tent.eq}
\title{Tent map: A coordinate transform of the logistic map}
\usage{
tent.eq(x)
}
\arguments{
\item{x}{last value of x, i.e. x[n]}
}
\description{
Qualitatively similar behavior to the logistic map,
but very visually clear/intuitive.
Use with iterate.map() function:
out <- iterate.map(a0=init.value, f=tent.eq, nSteps=num.of.iterations)
}
\examples{
tent.eq(.3)
}
\keyword{tent.map}
