% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-to-value-functions.R
\name{time2value}
\alias{time2value}
\title{Number of time periods to a specific value}
\usage{
time2value(x0, a, value)
}
\arguments{
\item{x0}{initial value}

\item{a}{multiplicative constant (factor of the geometric progression)}

\item{value}{value to be attained}
}
\description{
How long does a linear dynamical system take to reach
a given value from a given initial condition?
}
\examples{
time2value(100,.5,50) #1
time2value(200,.5,50) #2
}
