% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-to-value-functions.R
\name{timeto}
\alias{timeto}
\title{Number of time periods for an affine dynamical system to be
within a given "error" of its equilibrium}
\usage{
timeto(model = NULL, a = NULL, b = NULL, y0, final.diff = 0.01)
}
\description{
Function calculates the equilibrium for an empirical or
theoretical affine model, and returns how many time periods it takes
to get within a small margin of the equilibrium (the argument "final.diff",
which defaults to .01).

Uses the absolute value of difference between starting and equilibrium
value to allow for increasing or decreasing (or oscillatory) convergence
to the equilibrium. The "error" system, e[n+1] = a*e[n], where
e[n] = y[n] - y.star, will
always decrease towards zero in an affine system.

Result returned should be rounded to nearest integer for interpretation!
}
