% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boolean-network-functions.R
\name{transition.graph}
\alias{transition.graph}
\title{Mask for igraph package's graph_from_adjacency_matrix}
\usage{
transition.graph(a, ...)
}
\description{
takes the output of the the make.adjacency function to produce a graph of the state transitions
for a given update rule on a set of variables. Intended for Boolean networks but will work with any
update function on a discrete and finite state space (though the graph could become unmanageably large.)
}
