% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{triangular.n}
\alias{triangular.n}
\title{function to calculate the Nth triangular number}
\usage{
triangular.n(N)
}
\arguments{
\item{N}{the magnitude of the triangular number you want to calculate}
}
\description{
Returns the Nth triangular number (the sum of the natural numbers to N)
}
\examples{
naturals <- 1:15
for (n  in naturals ) {
  cat(triangular.n(n), "\n")
 }
 @export
}
