% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete-dynamic-utilities.R
\name{updown}
\alias{updown}
\title{Calculate the "Up-Down" signature of a vector}
\usage{
updown(x)
}
\arguments{
\item{x}{A numeric vector}
}
\description{
The up-down signature of a vector (or sequence) is
the sign of the lag-1 forward differences of the vector
updown = sign(diff(vector))
}
\examples{
set.seed(12345)
x <- rnorm(10)
updown(x)
}
