% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculus.R
\name{zombies}
\alias{zombies}
\title{Simulate a zombie apocalypse}
\usage{
zombies(
  z0 = 1,
  alpha = 0.005,
  beta = 0.0095,
  zeta = 1e-04,
  delta = 1e-04,
  stime = 5,
  dt = 0.01,
  N = 500
)
}
\arguments{
\item{z0}{- initial number of zombie infectees (at time 0), defaults to 1}

\item{alpha}{-  "zombie destruction" parm, defaults to .005}

\item{beta}{- "new zombie' parm, defaults to .0095}

\item{zeta}{- "zombie resurrection" rate, defaults to .0001}

\item{delta}{- background death rate, defaults to .0001}

\item{stime}{- stopping time, defaults to 5 (days)}

\item{dt}{- time step for numerical solution by Euler method, defaults to .01}

\item{N}{- human population at time 0, defaults to 500, considered 1000s}
}
\description{
based on Munz, Hudea, Imad, & Smith? When zombies attack!
Solve a system of ODEs, based on MATLAB code by Philip Munz (uses
Euler's method).
dx/dt = lim deltax/deltat as deltat -> 0
dx/dt = f(x) ~ deltax/deltat = f(x), for small enough deltat
deltax = deltat*f(x)
deltax = x[t+1] - x[t], so
x[t+1] = x[t] + dt*f(x):
Additionally plots output
adapted for R by Seth Spain, 15 MAR 2016
}
\examples{
zout <- zombies()
}
